Get DeltaCore to successfully build on Linux
Replaced Apple-only image init with a custom one that uses swift-png, use
OpenCombine on linux, and import CoreFoundation where used.

+++ Sources/Core/Sources/Network/SocketOption.swift
@@ -127,7 +127,7 @@ public struct MembershipRequestSocketOption: SettableSocketOption {
   }
 
   public var level: Int32 {
-    IPPROTO_IP
+    Int32(IPPROTO_IP)
   }
 
   public func makeSocketValue(from value: MembershipRequest) -> MembershipRequest {
